pyproject.toml
[project]
name = "summertime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-api-python-client>=2.163.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]
